Copy dot-files
Dot-files are files that starts with a dot, so a dot will be prepended
to filename if the file being copied does not have one.

No, `*' will not be expanded to hidden files (dot-files) in shell, so
let's set a rule: only dot-files without a starting dot can be copied.
In this way, a new benefit emerged: one won't have to be surprised to
find (even worse, after a week) that his `.gitconfig' was overwritten by
another file which won't show up with `ls' or `ll'.

For files in home directory that does not start with a dot, they can be
classified to a new category `home-files', but currently I haven't see
any such file.

File `.vimrc' starts with a dot but it's classified to vim-files.